Remove `make` and update test project
It does not look likely anymore that we are going to updgrade to any
sort of "build" (besides copying files) system anytime soon, so this
commit removes the `Makefile` and with the nuisance of having to run
`make` everytime something somewhere changes.

Also updates the test projects `latexmkrc` to no longer point at `build`
(accordingly) and adds a `\duedate` to `test.tex` to make it compile
again.

The README has also been updated.