Add `bsteffan-lecturenotes` class
Adds a new class for math lecture notes. Based on `scrartcl`, it
comprises a more-or-less complete setup for complete typed lecture notes
including a suitable title page, lecture tracking, index and
bibliography support, colored theorem environments, you name it. It's in
a WIP-but-usable state, and since I am going to immediately start using
it, expect more work to trickle in in the next days to weeks.

As usual, the theorem setup and barebones facilities for customizing
`tikzcd` live in their own packages, `bsteffan-notethms` and
`bsteffan-notediagrams`, respectively. All three files make use of
`etoolbox` early, since some package (presumably Koma) is going to load
it anyways, so we get to use its improved toggle facilities over old
LaTeX core `\newif`s. I will change this in the other packages little by
little whenever I have to make changes since there really is no good
reason to be proactive here.

Finally, we also fix a word in the README (`bsteffan-breakoutthms` was
labeled a "class", not a "package").