`bsteffan-lecturenotes`: Expand & improve color setup
Changes some of the colors (throwing out colorblind friendliness in the
process (oof) because it is hard, I'm weak, discernibility of all colors
is not a strict requirement for this usecase, and the status quo is that
none of the scripts I'm writing right now will be read by people with
color blindness, to the best of my knowledge) and adds more, meaningful
color aliases. Also moves this part up to be before the titlepage setup
so that we can use a color in the title.

Also includes some small changes: `bsteffan-notethms` now includes an
`example` and a `note` environment, and `bsteffan-notediagrams` now
loads the `calc` TikZ library.