`bsteffan-lecturenotes`: replace `marginnote` with `notecolumn`
Koma has a (semi-experimental) package `scrlayer-notecolumn` that
implements margin notes via dedicated page layout mechanisms and pdf
magic. One of its purposes is to fix the (quite serious) shortfalls of
the older `marginpar` and to improve on alternative solutions such as
`marginnote`. Since we are putting lecture counters in the margins,
migrating to this seems like a good idea (especially since the migration
is so painless, cf. the changes in this commit).

We use the opportunity to also upgrade the way lecture info is formatted
in the margins, switching out a hand-rolled `tikz` box with `mdframed`
which we are loading anyways. In the process, we also define a proper
Koma font (`lectureblock`) for this.

Finally, we also reach down into the entrails of `hyperref` and patch it
so that footnote marks can be colored separately from ordinary links,
all in order to have them black (instead of orange, which does stand out
in a bad way).