`math` & `topology`: Add more math groups
Long overdue update which wraps the definitions of a whole bunch of
commands with another pair of `{}` so that you can, say, write `F_\Z`
directly instead of `F_{\Z}` which was so far necessary.

This commit is sort of quick-and-dirty and I am sure to have missed some
instances where this would be beneficial; these will be patched whenever
they appear in a use case.