Remove superfluous `bsteffan-theorems` package
This package seems to have snuck in when during part 1 of the move to
the new base class system, probably because I was unsure whether I
wanted the theorem handling part to be a part of the base packages or
its own thing.

Since I seem to have gone with the former, this class is nothing but a
vestige and therefore can go.