Add `bsteffan-notes` class
Adds a class to deal with "regular," i.e. non-lecture notes. The setup
is basically a slimmed-down version of `bsteffan-lecturenotes` with
modified title page.

We also move some things around, notable moving the test for
`bsteffan-lecturenotes` to `lecturenotes_test.tex` and adding a new test
`notes_test.tex` for the new class.

We also make some adjustment in internal macro names in
`bsteffan-lecturenotes` to avoid clashes with the new class.