Switch from TeX Gyre Pagella to XITS
I've been entertaining the thought of switching typefaces for a while
now, in particular because the script font of TeX Gyre Pagella is pretty
much terrible. I'm not exactly in love with XITS, but I tested out a few
options and it seems reasonable. I'll stick with it for a bit and see
how I feel about it.